% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elide_year.R
\name{elide_year}
\alias{elide_year}
\alias{elide_years}
\title{elide_year}
\usage{
elide_year(x, ...)
}
\arguments{
\item{x}{`RY`, `PY`, `CY`, or `BY` object}

\item{verbose}{logical}
}
\description{
To elide distinctions between "CY", "BY", "RY", etc.
Drops the timeline (that is, prefix) and returns a simple integer.
}
\note{
`elide_years()` is a backwards-compatible alias for `elide_year()`.
}
