% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_years.R
\name{expand_years}
\alias{expand_years}
\title{expand_years}
\usage{
expand_years(
  input_data,
  years,
  year_var = "year",
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{tabular data}

\item{years}{vector of class \code{YYYY} that you want to expand across}

\item{year_var}{character}

\item{verbose}{display messages}
}
\description{
expand_years
}
