% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{format.YYYY}
\alias{format.YYYY}
\title{#' @param x `YYYY` object
#' @param ... ignored
#'
#' @noRd
#' @export
as.integer.YYYY <- function (x, ...) {
  field(x, "year")
}}
\usage{
\method{format}{YYYY}(x, ...)
}
\description{
#' @param x `YYYY` object
#' @param tz character
#' @param ... ignored
#'
#' @noRd
#' @export
as.Date.YYYY <- function (x, ..., tz = "") {
  datestamp <- stringr::str_c(as.integer(x), "-01-01")
  as.Date(datestamp, tz = tz, ...)
}
}
\details{
#' @param x `YYYY` object
#' @param tz character
#' @param ... further arguments to [as.POSIXct()]
#'
#' @noRd
#' @export
as.POSIXct.YYYY <- function (x, tz = "", ...) {
  dttm <- ISOdatetime(as.integer(x), 01, 01, 00, 00, 00, tz = tz)
  as.POSIXct(dttm, tz = tz, ...)
}

#' @noRd
#' @export
as.double.YYYY <- function (x, ...) {
  as.double(as.integer(x, ...))
}
}
