% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_years.R
\name{gather_years}
\alias{gather_years}
\title{gather_years}
\usage{
gather_years(input_data, value_col = "ems_qty", na.rm = FALSE, ...)
}
\arguments{
\item{input_data}{\code{data.frame} with columns of values named 1990, 2000, etc. (four-digit years)}

\item{value_col}{name of column for gathered values}

\item{...}{currently ignored}
}
\value{
\code{data.frame} with columns \code{year} (int) and \code{ems_qty} (optional: change name via \code{value}) (dbl)
}
\description{
gather_years
}
