% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndays_in.R
\name{ndays_in}
\alias{ndays_in}
\title{Number of days in a year, including leap days}
\usage{
ndays_in(year)
}
\arguments{
\item{year}{four-digit year (character, numeric, or integer)}
}
\value{
integer
}
\description{
Number of days in a year, including leap days
}
