% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{new_YYYY}
\alias{new_YYYY}
\title{Only do this if x is non-empty; otherwise the result will be
a length-1 vector (it'll just be "RY" or whatever `timeline` is)}
\usage{
new_YYYY(x = character(), timeline = NULL, verbose = TRUE)
}
\description{
Only do this if x is non-empty; otherwise the result will be
a length-1 vector (it'll just be "RY" or whatever `timeline` is)
}
