% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY.R
\name{[.BY}
\alias{[.BY}
\title{----------------------------------------------------------------------}
\usage{
\method{[}{BY}(x, i)
}
\description{
----------------------------------------------------------------------
}
