% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CY.R
\name{[.CY}
\alias{[.CY}
\title{----------------------------------------------------------------------}
\usage{
\method{[}{CY}(x, i)
}
\description{
----------------------------------------------------------------------
}
