% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PY.R
\name{[.PY}
\alias{[.PY}
\title{----------------------------------------------------------------------}
\usage{
\method{[}{PY}(x, i)
}
\description{
----------------------------------------------------------------------
}
