% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RY.R
\name{[.RY}
\alias{[.RY}
\title{----------------------------------------------------------------------}
\usage{
\method{[}{RY}(x, i)
}
\description{
----------------------------------------------------------------------
}
