% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{[.YYYY}
\alias{[.YYYY}
\title{----------------------------------------------------------------------}
\usage{
\method{[}{YYYY}(x, i)
}
\arguments{
\item{x}{`YYYY` object}

\item{i}{integer}
}
\description{
----------------------------------------------------------------------
}
