% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline}
\alias{timeline}
\alias{timeline.default}
\title{timeline}
\usage{
timeline(x)

\method{timeline}{default}(x)
}
\arguments{
\item{x}{[YYYY] object}

\item{pattern}{`x` needs to match this}
}
\description{
Get the timeline ("RY", "CY", etc.) associated with a `YYYY` object.
}
