% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_years.R
\name{unpack_years}
\alias{unpack_years}
\title{unpack_years}
\usage{
unpack_years(input_data, year_var = NULL, verbose = getOption("verbose"))
}
\arguments{
\item{input_data}{tabular data}

\item{year_var}{character or \code{NULL}}

\item{verbose}{display messages}
}
\description{
Unpack a vector of "packed years" (like RY or PY) inside a tibble
}
\note{
Might be slow and need optimization
}
\examples{
input_data <- data.frame(year = c("RY(2007:2009)", "RY(2001)"), ems_qty = c(111, 22))
unpack_years(input_data, year_var = "year")
unpack_years(input_data)

}
\seealso{
\code{\link[=RY]{RY()}} \code{\link[=PY]{PY()}}
}
