% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{vec_ptype2.YYYY.YYYY}
\alias{vec_ptype2.YYYY.YYYY}
\title{#' @export
print.YYYY <- function (x, ...) {
  vec_print_header.YYYY(x)
  if (length(x) == 0) return()
  out <- stats::setNames(format(x), names(x))
  print(out, quote = FALSE)
}}
\usage{
\method{vec_ptype2}{YYYY.YYYY}(x, y, ...)
}
\description{
#' @export
vec_print_header.YYYY <- function (x, ...) {
  cli::cat_line("<", attr(x, "timeline"), "[", length(x), "]>")
}
}
\details{
#' @noRd
vec_ptype_abbr.YYYY <- function (x, ...) {
  attr(x, "timeline")
}
}
