% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY-class.R, R/CY-class.R, R/PY-class.R,
%   R/RY-class.R, R/YYYY-class.R
\name{BY}
\alias{BY}
\alias{CY}
\alias{PY}
\alias{RY}
\alias{years}
\title{Years (the DataBank way)}
\usage{
BY(...)

CY(...)

PY(...)

RY(...)
}
\arguments{
\item{yyyy}{four-digit integer(s)}
}
\description{
This family of functions helps to index requests for DataBank "files".
}
\section{Functions}{
\itemize{
\item \code{BY}: Base year(s)

\item \code{CY}: Calendar year(s)

\item \code{PY}: Permit year(s)

\item \code{RY}: Reporting year(s)
}}

\note{
**FIXME: more and better description.** Why do we need these? What are they for?
}
\examples{
# The constructors `RY()`, `PY()`, etc. all yield instances of a very simple "S3" class (`YYYY`).
PY(2011:2014) \%>\% inherits("YYYY")
PY(2011:2014) \%>\% as.character()
PY(2011:2014) \%>\% as.integer()
PY(2011:2014) - 1 # can add and subtract
PY(2011:2014) \%>\% max()

# Reporting year(s)
RY(2015)
RY(2012:2016)

}
\seealso{
[DB_XXXX_CONCORDANCE]
}
