% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY-arith.R
\name{YYYY-subtraction}
\alias{YYYY-subtraction}
\alias{-.YYYY}
\title{Addition and subtraction}
\usage{
\method{-}{YYYY}(e1, e2)
}
\arguments{
\item{e1}{\code{YYYY} object}

\item{e2}{\code{YYYY} object}
}
\description{
Addition and subtraction
}
\note{
supports regression and prediction use-cases
}
