% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{YYYY}
\alias{YYYY}
\alias{RY}
\alias{PY}
\alias{BY}
\alias{CY}
\title{Typically we should call `YYYY()` rather than `new_YYYY()`.}
\usage{
YYYY(
  x = character(),
  ...,
  prefix = NULL,
  pattern = "^([CRPB]Y)?([0-9]{4})$",
  verbose = getOption("verbose", default = FALSE)
)

RY(x = character(), ...)

PY(x = character(), ...)

BY(x = character(), ...)

CY(x = character(), ...)
}
\arguments{
\item{x}{four-character year, like "2014"}

\item{...}{concatenated with `x`}

\item{prefix}{two-character prefix, like "RY" or "CY"}

\item{pattern}{`x` (and `...`, if supplied) need to match this}

\item{verbose}{logical}
}
\description{
Typically we should call `YYYY()` rather than `new_YYYY()`.
}
\section{Functions}{
\itemize{
\item \code{RY}: shortcut for `YYYY(..., prefix = "RY")`

\item \code{PY}: shortcut for `YYYY(..., prefix = "PY")`

\item \code{BY}: shortcut for `YYYY(..., prefix = "BY")`

\item \code{CY}: shortcut for `YYYY(..., prefix = "CY")`
}}

