% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elide_year.R
\name{elide_year}
\alias{elide_year}
\alias{elide_years}
\title{elide_year}
\usage{
elide_year(x, ...)

elide_year(x, ...)
}
\arguments{
\item{x}{`CY`, `BY`, `RY`, or `PY` object}

\item{...}{futher arguments}

\item{verbose}{logical}
}
\description{
To elide distinctions between "CY", "BY", "RY", etc.
Drops the timeline (that is, prefix) and returns a simple integer.

To elide distinctions between "CY", "BY", "RY", etc.
}
