% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elide_year.R
\name{elide_year}
\alias{elide_year}
\title{elide_year}
\usage{
elide_year(x, verbose = getOption("verbose"))
}
\arguments{
\item{x}{`RY`, `PY`, `CY`, or `BY` object}

\item{verbose}{display messages}
}
\description{
To elide distinctions between "CY", "BY", "RY", etc.
}
