% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elide_year.R
\name{elide_year}
\alias{elide_year}
\title{elide_year}
\usage{
elide_year(x, pattern = "([CRPB]Y)?([0-9]{4})", verbose = getOption("verbose"))
}
\arguments{
\item{x}{\code{RY}, \code{PY}, \code{CY}, or \code{BY} object}

\item{verbose}{logical}
}
\description{
To elide distinctions between "CY", "BY", "RY", etc.
Drops the timeline (that is, prefix) and returns a simple integer.
}
