% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elide_year.R, R/elide_years.R
\name{elide_year.default}
\alias{elide_year.default}
\alias{elide_year.data.frame}
\alias{elide_years}
\alias{elide_years.default}
\alias{elide_years.data.frame}
\title{elide_years}
\usage{
\method{elide_years}{default}(
  x,
  pattern = "^([CRPB]Y)?([0-9]{4})$",
  verbose = getOption("verbose")
)

\method{elide_years}{data.frame}(x, ...)

elide_years(x, ...)

\method{elide_years}{default}(x, ...)

\method{elide_years}{data.frame}(x, ...)
}
\arguments{
\item{x}{`CY`, `BY`, `RY`, or `PY` object}

\item{...}{futher arguments}
}
\description{
To elide distinctions between "CY", "BY", "RY", etc.
}
