% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY-format.R
\name{format.YYYY}
\alias{format.YYYY}
\title{#' @param x `YYYY` object
#' @param ... ignored
#'
#' @noRd
#' @export
format.YYYY <- function (x, ...) {
  attr(x, "timeline") <- NULL
  return(unclass(x))
}}
\usage{
\method{format}{YYYY}(x, ...)
}
\description{
#' @param x `YYYY` object
#' @param ... ignored
#' @noRd
#'
#' @export
print.YYYY <- function (x, ...) {
  x <- as.character(x)
  NextMethod()
}
#' @export
format.YYYY <- function(x, ...) {
  x_valid <- which(!is.na(x))
  year <- field(x, "year")[x_valid]
  ret <- rep(NA_character_, vec_size(x))
  ret[x_valid] <- paste0(timeline(x), format(year))
  return(ret)
}
}
