% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY-format.R
\name{format.YYYY}
\alias{format.YYYY}
\title{#' @param x `YYYY` object
#' @param ... ignored
#'
#' @noRd
#' @export
format.YYYY <- function (x, ...) {
  attr(x, "timeline") <- NULL
  return(unclass(x))
}}
\usage{
\method{format}{YYYY}(x, ...)
}
\description{
#' @param x `YYYY` object
#' @param ... ignored
#' @noRd
#'
#' @export
print.YYYY <- function (x, ...) {
  x <- as.character(x)
  NextMethod()
}
}
