% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_years.R
\name{gather_years}
\alias{gather_years}
\title{gather_years}
\usage{
gather_years(
  input_data,
  value_var,
  year_var = "year",
  pattern = "([A-Z]Y)([[:digit:]]{4})",
  na.rm = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{(tabular data)}

\item{value_var}{(character or symbol) defaults to "ems_qty"}

\item{year_var}{(character)}

\item{pattern}{(regexp) gather columns matching this}

\item{na.rm}{(logical) drop rows from output where value would be \code{NA}}

\item{verbose}{(logical) display messages}
}
\value{
tabular data with column \code{year}
}
\description{
Gather columns beginning with \code{CY}, \code{RY}, \code{PY}, as well as columns that look like "naive" four-digit years.
}
