% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_years.R
\name{gather_years}
\alias{gather_years}
\title{gather_years}
\usage{
gather_years(
  input_data,
  ...,
  year_var = "year",
  year_pattern = "([A-Z]Y)?[[:digit:]]{4}",
  convert = TRUE,
  na.rm = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{tabular data}

\item{...}{name of column for gathered values; defaults to `ems_qty`}

\item{convert}{as in [tidyr::gather()] (runs [type.convert()])}

\item{na.rm}{as in [tidyr::gather()] (drops rows from output where value would be `NA`)}

\item{verbose}{display messages}

\item{pattern}{gather columns matching this regexp}
}
\value{
tabular data with column `year`
}
\description{
Gather columns beginning with `CY`, `RY`, `PY`, as well as columns that look like "naive" four-digit years.
}
