% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{new_YYYY}
\alias{new_YYYY}
\title{new_YYYY}
\usage{
new_YYYY(x = character(), timeline = NULL, verbose = TRUE)
}
\arguments{
\item{x}{four-character year, like "2014"}

\item{timeline}{two-character prefix, like "RY" or "CY"}

\item{verbose}{logical}
}
\description{
new_YYYY
}
