% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_year.R
\name{parse_year}
\alias{parse_year}
\title{parse_year}
\usage{
parse_year(x, ...)
}
\arguments{
\item{x}{character, numeric, or YYYY-class}

\item{...}{ignored for now}
}
\value{
four-digit integer
}
\description{
Extracts a four-digit "naive" representation from various other representations of a "year".
}
\examples{
parse_year(CY(2030))
parse_year(RY(2015))
parse_year(BY(2011))
parse_year("1999")

}
