% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_years.R
\name{spread_years}
\alias{spread_years}
\title{spread_years}
\usage{
spread_years(
  input_data,
  value_var,
  year_var,
  ...,
  verbose = getOption("verbose")
)
}
\arguments{
\item{input_data}{(tabular data)}

\item{value_var}{(character or symbol) defaults to whatever is returned by \code{\link[=find_qty_var]{find_qty_var()}}.}

\item{year_var}{(character) defaults to whatever is found by \code{\link[=find_year_var]{find_year_var()}}.}

\item{...}{(dots) passed to \code{\link[tidyr:spread]{tidyr::spread()}}}

\item{verbose}{(logical) display messages}
}
\value{
tabular data with column \code{year}
}
\description{
Spread \code{year} into columns.
}
