% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY-timeline.R, R/timeline.R
\name{timeline}
\alias{timeline}
\title{timeline}
\usage{
timeline(x, pattern = "^([CRPB]Y)?([0-9]{4})$")

timeline(x, pattern = "^([CRPB]Y)?([0-9]{4})$")
}
\arguments{
\item{x}{[YYYY] object}

\item{pattern}{`x` needs to match this}
}
\description{
Get the timeline ("CY", "RY", etc.) of a `YYYY` object.

Get the timeline ("RY", "CY", etc.) associated with a `YYYY` object.
}
