% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_years.R
\name{unpack_years}
\alias{unpack_years}
\title{Unpack a vector of "packed years" (like RY or PY)}
\usage{
unpack_years(input_data, year_var = NULL,
  verbose = getOption("verbose"))
}
\description{
Unpack a vector of "packed years" (like RY or PY)
}
\note{
Might be slow and need optimization
}
\examples{
input_data <- tibble(years = c("RY(2007:2009)", "RY(2001)", "PY(2003)"), ems_qty = c(111, 22, 45))
unpack_years(input_data, year_var = "years")
unpack_years(input_data)

}
\seealso{
[RY()] [PY()]
}
