% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R, R/BY.R, R/CY.R, R/PY.R, R/RY.R
\name{new_YYYY}
\alias{new_YYYY}
\alias{BY}
\alias{CY}
\alias{PY}
\alias{RY}
\title{YYYY}
\usage{
new_YYYY(..., timeline = c("CY", "RY", "PY", "BY"))

BY(...)

CY(x)

PY(...)

RY(...)
}
\arguments{
\item{...}{integer(s)}

\item{prefix}{character}
}
\description{
This family of functions helps to index requests for DataBank "files".
}
\section{Functions}{
\itemize{
\item \code{BY}: Base year(s)

\item \code{CY}: Calendar year(s)

\item \code{PY}: Permit year(s)

\item \code{RY}: Reporting year(s)
}}

\note{
**FIXME: more and better description.** Why do we need these? What are they for?
}
\examples{
# The constructors `RY()`, `PY()`, etc. all yield instances of a very simple "S3" class (`YYYY`).
CY(2011:2014) \%>\% inherits("YYYY")
PY(2011:2014) \%>\% as.character()
BY(2011:2014) \%>\% as.integer()
CY(2011:2014) \%>\% max()

}
\seealso{
[DB_XXXX_CONCORDANCE]
}
\concept{YYYY}
