% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YYYY.R
\name{YYYY}
\alias{YYYY}
\alias{new_YYYY}
\alias{CY}
\alias{RY}
\alias{PY}
\alias{BY}
\title{Years with explicit timelines}
\usage{
new_YYYY(..., timeline = c("CY", "RY", "PY", "BY"))

CY(...)

RY(...)

PY(...)

BY(...)
}
\arguments{
\item{...}{positive whole numbers}

\item{timeline}{character}
}
\description{
This family of functions helps to index requests for DataBank "files",
and helps with the construction of "base year" emission inventories.

The constructors \code{RY()}, \code{PY()}, etc. all yield instances of a very simple "S3" class (\code{YYYY}).
}
\section{Functions}{
\itemize{
\item \code{CY}: Calendar year(s)

\item \code{RY}: Reporting year(s)

\item \code{PY}: Permit year(s)

\item \code{BY}: Base year(s)
}}

\references{
\itemize{
\item \href{https://paper.dropbox.com/doc/Key-Concept-Whats-in-a-Year--BEm7K0si5VGV_nTugk5BHqo0Ag-p6bICw8OZNDcUazyCFULw}{What's In a Year?}
}
}
